chore: fix circular dependency error
