+++ src/modules/notifications/dtos/mark-all-notifications-as-read-error.dto.ts
@@ -1,0 +1,6 @@
+export class MarkAllNotificationAsReadError {
+  status: boolean;
+  status_code: number;
+  error: string;
+  message: string;
+}
