+++ src/modules/notifications/dtos/mark-all-notifications-as-read.dto.ts
@@ -1,17 +1,8 @@
-export class MarkAllNotificationAsReadError {
-  status: boolean;
-  status_code: number;
-  error: string;
-  message: string;
-}
-
-class Data {
-  notifications: [];
-}
-
 export class MarkAllNotificationAsReadResponse {
   status: string;
   status_code: number;
   message: string;
-  data: Data;
+  data: {
+    notifications: [];
+  };
 }
